[SAI-PTF]Reformat PTF brcm port cofig with Port and Port_config object Why Organize port-related data and port-related config into objects Cause there are multi ASICs and versions, for differences of them, there might be different APIs, and some of them can be reused of them just for certain versions or in special API sequences. In order to maintain them better, then put them into a centra place, and plan to organize them with same manner in SAI, within the object scope.
> this PR depends on https://github.com/opencomputeproject/SAI/pull/1711

How
Add data_module and config for the data object and config class for a certain data object.
reformat code, reorganize the code with the data module and config class
> Current data module and data object also compatible with existing data module(lists in sai_test_base)

Test
Duts with cases
```
ptf --test-dir ptf saisanity.L2SanityTest --relax --xunit --xunit-dir "/tmp/sai_qualify/test_results_tmp" --interface '0-0@eth0' --interface '0-1@eth1' --interface '0-2@eth2' --interface '0-3@eth3' --interface '0-4@eth4' --interface '0-5@eth5' --interface '0-6@eth6' --interface '0-7@eth7' --interface '0-8@eth8' --interface '0-9@eth9' --interface '0-10@eth10' --interface '0-11@eth11' --interface '0-12@eth12' --interface '0-13@eth13' --interface '0-14@eth14' --interface '0-15@eth15' --interface '0-16@eth16' --interface '0-17@eth17' --interface '0-18@eth18' --interface '0-19@eth19' --interface '0-20@eth20' --interface '0-21@eth21' --interface '0-22@eth22' --interface '0-23@eth23' --interface '0-24@eth24' --interface '0-25@eth25' --interface '0-26@eth26' --interface '0-27@eth27' --interface '0-28@eth28' --interface '0-29@eth29' --interface '0-30@eth30' --interface '0-31@eth31' --relax "--test-params=thrift_server='$DUTIP';port_config_ini='/tmp/sai_qualify/sai_test/resources/port_config.ini'"

ptf --test-dir test/sai_test sai_sanity_test.SaiSanityTest --relax --xunit --xunit-dir "/tmp/sai_qualify/test_results_tmp" --interface '0-0@eth0' --interface '0-1@eth1' --interface '0-2@eth2' --interface '0-3@eth3' --interface '0-4@eth4' --interface '0-5@eth5' --interface '0-6@eth6' --interface '0-7@eth7' --interface '0-8@eth8' --interface '0-9@eth9' --interface '0-10@eth10' --interface '0-11@eth11' --interface '0-12@eth12' --interface '0-13@eth13' --interface '0-14@eth14' --interface '0-15@eth15' --interface '0-16@eth16' --interface '0-17@eth17' --interface '0-18@eth18' --interface '0-19@eth19' --interface '0-20@eth20' --interface '0-21@eth21' --interface '0-22@eth22' --interface '0-23@eth23' --interface '0-24@eth24' --interface '0-25@eth25' --interface '0-26@eth26' --interface '0-27@eth27' --interface '0-28@eth28' --interface '0-29@eth29' --interface '0-30@eth30' --interface '0-31@eth31' --relax "--test-params=thrift_server='$DUTIP';port_config_ini='/tmp/sai_qualify/sai_test/resources/port_config.ini'"
```
