+++ ptf/data_module/data_obj.py
@@ -1,0 +1,54 @@
+# Copyright (c) 2021 Microsoft Open Technologies, Inc.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+#    THIS CODE IS PROVIDED ON AN *AS IS* BASIS, WITHOUT WARRANTIES OR
+#    CONDITIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT
+#    LIMITATION ANY IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS
+#    FOR A PARTICULAR PURPOSE, MERCHANTABILITY OR NON-INFRINGEMENT.
+#
+#    See the Apache Version 2.0 License for specific language governing
+#    permissions and limitations under the License.
+#
+#    Microsoft would like to thank the following companies for their review and
+#    assistance with these files: Intel Corporation, Mellanox Technologies Ltd,
+#    Dell Products, L.P., Facebook, Inc., Marvell International Ltd.
+#
+#
+
+
+def auto_str(cls):
+    """
+    class decorator for auto generate the __str__ method for all the class attributes.
+    """
+    def __str__(self):
+        """
+        auto generate __str__ for all the class atrributes.
+        """
+        return '%s(%s)' % (
+            type(self).__name__,
+            ', '.join('%s=%s' % item for item in vars(self).items())
+        )
+    cls.__str__ = __str__
+    return cls
+
+@auto_str
+class data_item():
+    """
+    Represent the basic data object.
+    Attrs:
+        oid: object id
+    """
+
+    def __init__(self, oid = None):
+        """
+        Init data item Object
+        Init following attrs:
+            oid: object id
+        """
+        self.oid = oid
+        """
+        object id
+        """
